`@remotion/effects`: Migrate starburst effect
