+++ packages/brand/package.json
@@ -28,7 +28,6 @@
 		"@remotion/rough-notation": "workspace:*",
 		"@remotion/shapes": "workspace:*",
 		"@remotion/sfx": "workspace:*",
-		"@remotion/starburst": "workspace:*",
 		"@remotion/svg-3d-engine": "workspace:*",
 		"@remotion/captions": "workspace:*",
 		"@remotion/tailwind-v4": "workspace:*",
