+++ packages/docs/components/demos/StarburstDemo.tsx
@@ -1,36 +1,31 @@
-import {Starburst} from '@remotion/starburst';
+import {starburst} from '@remotion/effects/starburst';
 import React from 'react';
-import {AbsoluteFill, useVideoConfig} from 'remotion';
+import {
+	AbsoluteFill,
+	interpolate,
+	Solid,
+	useCurrentFrame,
+	useVideoConfig,
+} from 'remotion';
 
-interface Props {
-	readonly rays: number;
-	readonly rotation: number;
-	readonly smoothness: number;
-	readonly vignette: number;
-	readonly originOffsetX: number;
-	readonly originOffsetY: number;
-}
+export const StarburstDemoComp: React.FC = () => {
+	const frame = useCurrentFrame();
+	const {width, height} = useVideoConfig();
 
-export const StarburstDemoComp: React.FC<Props> = ({
-	rays,
-	rotation,
-	smoothness,
-	vignette,
-	originOffsetX,
-	originOffsetY,
-}) => {
-	const {durationInFrames} = useVideoConfig();
 	return (
 		<AbsoluteFill style={{backgroundColor: 'black'}}>
-			<Starburst
-				rays={rays}
-				colors={['#ffdd00', '#ff8800']}
-				rotation={rotation}
-				smoothness={smoothness}
-				vignette={vignette}
-				originOffsetX={originOffsetX}
-				originOffsetY={originOffsetY}
-				durationInFrames={durationInFrames}
+			<Solid
+				width={width}
+				height={height}
+				effects={[
+					starburst({
+						rays: 16,
+						colors: ['#ffdd00', '#ff8800'],
+						rotation: interpolate(frame, [0, 45, 89], [0, 180, 360]),
+						smoothness: interpolate(frame, [0, 45, 89], [0, 0.3, 0]),
+						origin: [0.5, 0.5],
+					}),
+				]}
 			/>
 		</AbsoluteFill>
 	);
