+++ packages/docs/components/demos/types.ts
@@ -1545,65 +1545,10 @@ export const starburstDemo: DemoType = {
 	durationInFrames: 90,
 	fps: 30,
 	id: 'starburst',
-	autoPlay: false,
+	autoPlay: true,
 	controls: false,
 	logLevel: 'info',
-	options: [
-		{
-			name: 'rays',
-			type: 'numeric',
-			default: 12,
-			min: 2,
-			max: 100,
-			step: 1,
-			optional: 'no',
-		},
-		{
-			name: 'rotation',
-			type: 'numeric',
-			default: 0,
-			min: 0,
-			max: 360,
-			step: 1,
-			optional: 'no',
-		},
-		{
-			name: 'smoothness',
-			type: 'numeric',
-			default: 0,
-			min: 0,
-			max: 1,
-			step: 0.01,
-			optional: 'no',
-		},
-		{
-			name: 'vignette',
-			type: 'numeric',
-			default: 1,
-			min: 0,
-			max: 1,
-			step: 0.01,
-			optional: 'no',
-		},
-		{
-			name: 'originOffsetX',
-			type: 'numeric',
-			default: 0,
-			min: -1,
-			max: 1,
-			step: 0.01,
-			optional: 'no',
-		},
-		{
-			name: 'originOffsetY',
-			type: 'numeric',
-			default: 0,
-			min: -1,
-			max: 1,
-			step: 0.01,
-			optional: 'no',
-		},
-	],
+	options: [],
 };
 
 export const transitionSeriesTransitionDemo: DemoType = {
