+++ packages/docs/docs/starburst/index.mdx
@@ -8,6 +8,10 @@ crumb: API
 
 # @remotion/starburst<AvailableFrom v="4.0.435" />
 
+:::warning Deprecated
+`@remotion/starburst` is deprecated and will no longer be published starting with Remotion 5.0. Migrate to [`starburst()`](/docs/effects/starburst) from `@remotion/effects/starburst`. The `<Starburst>` component will not be moved to `@remotion/effects`.
+:::
+
 Starburst ray effect for Remotion videos.
 
 import {TableOfContents} from './table-of-contents';
