+++ packages/docs/docs/starburst/starburst-effect.mdx
@@ -8,6 +8,10 @@ crumb: '@remotion/starburst'
 
 # starburst()<AvailableFrom v="4.0.468" />
 
+:::warning Deprecated
+`@remotion/starburst` is deprecated and will no longer be published starting with Remotion 5.0. Migrate to [`starburst()`](/docs/effects/starburst) from `@remotion/effects/starburst`.
+:::
+
 _Part of the [`@remotion/starburst`](/docs/starburst/api) package._
 
 Renders a WebGL2-based retro starburst ray pattern as an effect for canvas-based components such as [`<Video>`](/docs/media/video), [`<HtmlInCanvas>`](/docs/remotion/html-in-canvas) and [`<Solid>`](/docs/solid).
@@ -83,6 +87,7 @@ When `true`, the effect is skipped. Defaults to `false`.
 ## See also
 
 - [`@remotion/starburst`](/docs/starburst/api)
+- [`starburst()` from `@remotion/effects`](/docs/effects/starburst)
 - [`<Starburst>`](/docs/starburst/starburst)
 - [`@remotion/effects`](/docs/effects/api)
 - [Source code for this effect](https://github.com/remotion-dev/remotion/blob/main/packages/starburst/src/starburst-effect.ts)
