+++ packages/docs/src/data/articles.ts
@@ -2223,6 +2223,15 @@ export const articles = [
 		slug: 'effects/speckle',
 	},
 	{
+		id: 'effects/starburst',
+		title: 'starburst()',
+		relativePath: 'docs/effects/starburst.mdx',
+		compId: 'articles-docs-effects-starburst',
+		crumb: '@remotion/effects',
+		noAi: false,
+		slug: 'effects/starburst',
+	},
+	{
 		id: 'effects/thermal-vision',
 		title: 'thermalVision()',
 		relativePath: 'docs/effects/thermal-vision.mdx',
@@ -6738,7 +6747,7 @@ export const articles = [
 		title: 'Starburst',
 		relativePath: 'docs/starburst-guide.mdx',
 		compId: 'articles-docs-starburst-guide',
-		crumb: 'Techniques',
+		crumb: 'Designing visuals',
 		noAi: false,
 		slug: 'starburst',
 	},
