+++ packages/effects/src/index.ts
@@ -1,6 +1,12 @@
 export {checkerboard, type CheckerboardParams} from './checkerboard.js';
 export {pattern, type PatternOrigin, type PatternParams} from './pattern.js';
 export {rings, type RingsCenter, type RingsParams} from './rings.js';
+export {
+	starburst,
+	starburstEffectSchema,
+	type StarburstEffectParams,
+	type StarburstOrigin,
+} from './starburst.js';
 export {gridlines, type GridlinesParams} from './gridlines.js';
 export {zigzag, type ZigzagDirection, type ZigzagParams} from './zigzag.js';
 export {
