+++ packages/skills/skills/remotion-markup/effects.md
@@ -18,7 +18,7 @@ Install the package that provides the chosen effect:
 npx remotion add @remotion/effects
 ```
 
-Use `npx remotion add @remotion/light-leaks` for `lightLeak()` and `npx remotion add @remotion/starburst` for `starburst()`.
+Use `npx remotion add @remotion/light-leaks` for `lightLeak()`.
 
 Effects are functions passed to the `effects` prop of canvas-based components such as `<Video>` from `@remotion/media`, `<Solid>`, `<CanvasImage>`, and `<HtmlInCanvas>`.
 
@@ -29,7 +29,7 @@ import {blur} from '@remotion/effects/blur';
 <Video src="https://remotion.media/video.mp4" effects={[blur({radius: 8})]} />;
 ```
 
-Use the effect docs for exact props and imports. Most `@remotion/effects` imports use `@remotion/effects/<effect-slug>`; `uvTranslate()` and `xyTranslate()` use `@remotion/effects/translate`; `lightLeak()` uses `@remotion/light-leaks`; `starburst()` uses `@remotion/starburst`.
+Use the effect docs for exact props and imports. Most `@remotion/effects` imports use `@remotion/effects/<effect-slug>`; `uvTranslate()` and `xyTranslate()` use `@remotion/effects/translate`; `lightLeak()` uses `@remotion/light-leaks`.
 
 These effects use WebGL2. During renders, enable WebGL with:
