+++ packages/studio-shared/src/effect-catalog.ts
@@ -18,10 +18,6 @@ export const getEffectDocumentationPath = (item: EffectCatalogItem) => {
 		return '/docs/light-leaks/light-leak-effect';
 	}
 
-	if (item.id === 'effects-starburst') {
-		return '/docs/starburst/starburst-effect';
-	}
-
 	return `/docs/effects/${item.id.slice('effects-'.length)}`;
 };
 
@@ -792,7 +788,7 @@ export const EFFECT_CATALOG: readonly EffectCatalogItem[] = [
 		description: 'Starburst ray effect',
 		effect: {
 			name: 'starburst',
-			importPath: '@remotion/starburst',
+			importPath: '@remotion/effects/starburst',
 			config: {
 				rays: 16,
 				colors: ['#ff6600', '#ffff00'],
