+++ packages/studio/src/test/effect-catalog.test.ts
@@ -37,6 +37,6 @@ test('effect catalog exposes documentation links for picker context menu', () =>
 		'https://www.remotion.dev/docs/light-leaks/light-leak-effect',
 	);
 	expect(links.get('starburst()')).toBe(
-		'https://www.remotion.dev/docs/starburst/starburst-effect',
+		'https://www.remotion.dev/docs/effects/starburst',
 	);
 });
